Разделил фабрику и фигуры
