+++ ViewModels/FigureFabric.cs
@@ -3,8 +3,7 @@ using System.Collections.Generic;
 using System.Composition;
 using System.Composition.Hosting;
 using System.Linq;
-using System.Text.Json.Serialization;
-using Avalonia;
+using Paint2.Models.Figures;
 using Interfaces;
 using Point = Interfaces.Point;
 
@@ -53,105 +52,5 @@ public static class FigureFabric
     {
         return info.AvailableFigures.First(f => f.Metadata.Name == FigureName).Value
             .Create(doubleParams,pointParams);
-    }
-}
-
-public class Circle : IFigure
-{
-    [Export(typeof(IFigureCreator))]
-    [ExportMetadata(nameof(FigureMetadata.Name), nameof(Circle))]
-    class CircleCreator : IFigureCreator
-    {
-        public int NumberOfDoubleParameters => 1;
-
-        public int NumberOfPointParameters => 1;
-
-        public IEnumerable<string> PointParametersNames
-        {
-            get
-            {
-                yield return "Center";
-            }
-        }
-
-        public IEnumerable<string> DoubleParametersNames
-        {
-            get
-            {
-                yield return "Radius";
-            }
-        }
-
-        public IFigure Create(IDictionary<string, double> doubleParams, IDictionary<string, Point> pointParams)
-        {
-            return new Circle(pointParams["Center"], doubleParams["Radius"]);
-        }
-    }
-    Point Center { get; set; }
-    double Radius { get; set; }
-    Circle(Point c,double r)
-    {
-        Center = c;
-        Radius = r;
-    }
-    public double Width => throw new NotImplementedException();
-
-    public double Height => throw new NotImplementedException();
-
-    public string Name => throw new NotImplementedException();
-
-    public void Render(IRenderInterface toDraw)
-    {
-        throw new NotImplementedException();
-    }
-
-    public IFigure Intersect(IFigure other)
-    {
-        throw new NotImplementedException();
-    }
-
-    public bool IsInternal(Point p, double eps)
-    {
-        throw new NotImplementedException();
-    }
-
-    public void Move(Point vector)
-    {
-        throw new NotImplementedException();
-    }
-
-    public void Reflection(Point ax1, Point ax2)
-    {
-        throw new NotImplementedException();
-    }
-
-    public void Rotate(Point Center, double angle)
-    {
-        throw new NotImplementedException();
-    }
-
-    public void Scale(double x, double y)
-    {
-        throw new NotImplementedException();
-    }
-
-    public void Scale(Point Center, double rad)
-    {
-        throw new NotImplementedException();
-    }
-
-    public void SetParameters(IDictionary<string, double> doubleParams, IDictionary<string, Point> pointParams)
-    {
-        throw new NotImplementedException();
-    }
-
-    public IFigure Subtract(IFigure other)
-    {
-        throw new NotImplementedException();
-    }
-
-    public IFigure Union(IFigure other)
-    {
-        throw new NotImplementedException();
     }
 }
